% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorramp_sig.R
\name{colorramp_sig}
\alias{colorramp_sig}
\title{Significance colour ramps}
\usage{
colorramp_sig(palette = "all", reverse = FALSE, random = FALSE, n = 0, ...)
}
\arguments{
\item{palette}{Name of the base palette. See \code{\link[=palette_sig]{palette_sig()}} for the
available options.}

\item{reverse}{Logical; if \code{TRUE}, reverse the colour order before
creating the colour ramp.}

\item{random}{Logical; if \code{TRUE}, randomly shuffle the colours before
creating the colour ramp.}

\item{n}{Number of colours to extract from the base palette before
constructing the colour ramp. Defaults to \code{0}, which uses all colours
in the selected palette.}

\item{...}{Additional arguments passed to \code{\link[grDevices:colorRamp]{grDevices::colorRampPalette()}}.}
}
\value{
A function generated by \code{\link[grDevices:colorRamp]{grDevices::colorRampPalette()}} that returns a
vector of colour hex codes when called with the desired number of colours.
}
\description{
Creates a colour ramp function based on one of the predefined
Significance colour palettes. The returned function can be used to
generate any number of interpolated colours and is particularly useful
for continuous colour scales in ggplot2.
}
\details{
Available palettes are identical to those supported by \code{\link[=palette_sig]{palette_sig()}}.
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(wt, mpg, colour = hp)) +
  geom_point(size = 3) +
  scale_colour_gradientn(
    colours = colorramp_sig("mintoplus")(100)
  ) +
  labs(
    colour = "Horsepower",
    x = "Weight",
    y = "Miles per gallon"
  )

}
\keyword{colour}
\keyword{palette}
