% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorramp_sig_function.R
\name{colorramp_sig_function}
\alias{colorramp_sig_function}
\title{Create colorramps for continuous scales}
\usage{
colorramp_sig_function(
  palette = "all",
  reverse = FALSE,
  random = FALSE,
  n = 0,
  ...
)
}
\value{
String or list of strings with color hex codes
}
\description{
Create colorramps for continuous scales
}
