% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_sig.R
\name{colors_sig}
\alias{colors_sig}
\title{Significance color codes}
\usage{
colors_sig(...)
}
\arguments{
\item{...}{String or list of strings with color name(s). Possible values are "light blue", "blue", "yellow", "orange", "red", "green", "light blue - darker", "blue - darker", "yellow - darker", "orange - darker", "red - darker", "green - darker", "light blue - lighter", "blue - lighter", "yellow - lighter", "orange - lighter", "red - lighter", "green - lighter", "very light blue", "lightest blue"}
}
\value{
String or list of strings with color hex codes
}
\description{
Definition of colors used in Significance style color palettes
}
\examples{
colors_sig("light blue")
colors_sig("blue", "light blue")
mtcars \%>\%
 ggplot2::ggplot(aes(gear)) +
   ggplot2::geom_bar(fill = colors_sig("blue")) +
   ggplot2::labs(x="Number of gears",
        y="Count",
        title="Car types"
   ) +
   theme_sig() +
   scale_y_continuous(expand = c(0, 0))

}
\keyword{color}
\keyword{palette}
