% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle-data.R
\docType{data}
\name{cycletrips}
\alias{cycletrips}
\title{Sample data with cycle trips}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 49805 rows and 14 columns.
}
\source{
ODiN 2018
}
\usage{
data(cycletrips)
}
\value{
Data set (tibble) of bicycle trips
}
\description{
Cycle trips (main trip mode only) from ODiN 2018
}
\examples{
data(cycletrips)
cycletrips \%>\% count(Weekdag)
}
\keyword{datasets}
