% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_sig_function.R
\name{palette_sig_function}
\alias{palette_sig_function}
\title{Create color palette for discrete scales}
\usage{
palette_sig_function(palette = "all", reverse = FALSE, random = FALSE)
}
\value{
String or list of strings with color hex codes
}
\description{
Create color palette for discrete scales
}
