% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sig.R
\name{theme_sig}
\alias{theme_sig}
\title{Significance ggplot2 theme}
\usage{
theme_sig(flip = FALSE, grid = FALSE)
}
\arguments{
\item{flip}{Logical; if \code{TRUE}, shows vertical grid lines. If \code{FALSE}
(default), shows horizontal grid lines. Only has effect if \code{grid = FALSE}.}

\item{grid}{Logical; if \code{TRUE}, shows both horizontal and vertical grid
lines. If \code{FALSE} (default), shows either horizontal or vertical grid
lines, depending on flip.}
}
\value{
Theme to be used for ggplot2 plots
}
\description{
Defines the theme for ggplot2 plots. This includes fonts, grid lines, axis
ticks, margins, etc.
}
\examples{
mtcars \%>\%
  ggplot(aes(gear)) +
    geom_bar() +
    labs(x="Number of gears",
         y="Count",
         title="Car types"
    ) +
    theme_sig() +
    scale_y_continuous(expand = c(0, 0))
mtcars \%>\%
  ggplot(aes(y=gear)) +
    geom_bar() +
    labs(x="Count",
         y="Number of gears",
         title="Car types"
    ) +
    theme_sig(flip = TRUE) +
    scale_x_continuous(expand = c(0, 0))

}
\keyword{theme}
