% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sig_map.R
\name{theme_sig_map}
\alias{theme_sig_map}
\title{Significance ggplot2 theme for maps}
\usage{
theme_sig_map()
}
\value{
Theme to be used for ggplot2 plots
}
\description{
Defines the theme for ggplot2 plots. This includes fonts, margins, etc.
}
\examples{
mtcars \%>\%
  ggplot(aes(gear)) +
    geom_bar() +
    labs(x="Number of gears",
         y="Count",
         title="Car types"
    ) +
    theme_sig_map()

}
\keyword{theme}
