% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_sig.R
\name{palette_sig}
\alias{palette_sig}
\title{Significance color palettes}
\usage{
palette_sig(palette = "all", n = 0)
}
\arguments{
\item{palette}{String with name of a color palette. Possible values are listed below.}

\item{n}{Number of colors to extract}
}
\value{
List with color names and hex values
}
\description{
Defines color palettes, e.g. for use in ggplot2 plots.
All color palettes are based on the colors from the Significance house style Excel template.
}
\details{
Currently available palettes are:
basic : the 2 colors from the Significance logo. Suitable for category or continuous data.
main : the standard set of 6 colors in the significance theme. Particularly suitable for category data.
lighter : the lighter 6 colors in significance theme. Particularly suitable for category data.
darker : the darker 6 colors in significance theme. Particularly suitable for category data.
all : all 18 colors in significance theme (main, lighter and darker combined). Particularly suitable for category data.
light : the lighter 12 colors in significance theme (main and lighter combined). Particularly suitable for category data.
posneg : green (positive) and red (negative). Particularly suitable for (binary) category data.
posnegneut : green (positive), red (negative) and light blue (neutral). Particularly suitable for category data.
postoneg : green (positive), white (neutral) and red (negative). Particularly suitable for continuous data.
mintoplus : blue (low), white (neutral) and red (high). Particularly suitable for continuous data.
blues : white (low) and blue (high). Particularly suitable for continuous data.
lightblues : white (low) and light blue (high). Particularly suitable for continuous data.
intense : yellow (low), orange (mid) and red (high). Particularly suitable for continuous data.
highlight : light blue (regular) and red (emphasis). Particularly suitable for highlighting a selection of data points.
intense : yellow (low), orange (mid) and red (high). Particularly suitable for continuous data.

For the color palettes "all", "main", "light", "lighter" and "darker" the colors are recycled if the required number of colors is larger than the number of colors in the palette. For the other palettes the colors will be interpolated.
}
\examples{
palette_sig("main")

}
\keyword{color}
\keyword{palette}
