% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_sig.R
\name{palette_sig}
\alias{palette_sig}
\title{Significance color palettes}
\usage{
palette_sig(palette = "all", n = 0)
}
\arguments{
\item{palette}{String with name of a color palette. Possible values are "all" (complete 18 colors palette), "main" (6 colors: light blue, blue, yellow, orange, red and green), "lighter" (same 6 colors, but lighter), "darker" (same 6 colors, but darker), "basic" (only the 2 Significance logo colors: light blue and blue), "posneg" (green and red) and "posnegneut" (green, red and light blue)}

\item{n}{Number of colors to extract}
}
\value{
List with color names and hex values
}
\description{
Defines color palettes, e.g. for use in ggplot2 plots.
}
\examples{
palette_sig("main")

}
\keyword{color}
\keyword{palette}
