% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_sig.R
\name{scale_fill_sig}
\alias{scale_fill_sig}
\title{Fill color scale ramps from Significance style color palettes}
\usage{
scale_fill_sig(
  palette = "all",
  discrete = TRUE,
  reverse = FALSE,
  random = FALSE,
  n = 0,
  ...
)
}
\arguments{
\item{palette}{String with name of a color palette. Possible values are "all" (complete 18 colors palette), "main" (6 colors: light blue, blue, yellow, orange, red and green), "lighter" (same 6 colors, but lighter), "darker" (same 6 colors, but darker), "basic" (only the 2 Significance logo colors: light blue and blue), "posneg" (green and red) and "posnegneut" (green, red and light blue)}

\item{discrete}{Indicates a discrete (TRUE) or continuous (FALSE) scale}

\item{reverse}{Reverses color ordering if TRUE}

\item{random}{Shuffles color ordering if TRUE}
}
\description{
Defines color scales for ggplot2 plots, based on color ramps from color values. For determining fill colors.
}
\examples{
mtcars \%>\%
mutate(gear = as_factor(gear)) \%>\%
  group_by(gear) \%>\%
  summarise(count = n()) \%>\%
  ggplot(aes(x="", y=count, fill=gear)) +
  geom_bar(stat="identity", width=1, color="white") +
  labs(title="Number of gears") +
  coord_polar("y", start=0, direction=-1) +
  theme_sig() +
  scale_fill_sig("basic", discrete=TRUE, reverse=FALSE) +
  scale_y_discrete(expand = c(0, 0))

}
\keyword{color}
\keyword{palette}
